fix(functions):アカウント処理用のファイルを作成・indexから移植
