+++ functions/src/index.ts
@@ -1,26 +1,4 @@
-import * as functions from 'firebase-functions';
 import * as admin from 'firebase-admin';
 export * from './post.function';
-
+export * from './user';
 admin.initializeApp();
-const db = admin.firestore();
-
-export const createUser = functions
-  .region('asia-northeast1')
-  .auth.user()
-  .onCreate((user) => {
-    return db.doc(`users/${user.uid}`).set({
-      uid: user.uid,
-      name: user.displayName,
-      avaterURL: user.photoURL,
-      email: user.email,
-      createdAt: new Date(),
-    });
-  });
-
-export const deleteUser = functions
-  .region('asia-northeast1')
-  .auth.user()
-  .onDelete((user) => {
-    return db.doc(`users/${user.uid}`).delete();
-  });
