+++ functions/src/user.ts
@@ -1,0 +1,24 @@
+import * as functions from 'firebase-functions';
+import * as admin from 'firebase-admin';
+
+const db = admin.firestore();
+
+export const createUser = functions
+  .region('asia-northeast1')
+  .auth.user()
+  .onCreate((user) => {
+    return db.doc(`users/${user.uid}`).set({
+      uid: user.uid,
+      name: user.displayName,
+      avaterURL: user.photoURL,
+      email: user.email,
+      createdAt: new Date(),
+    });
+  });
+
+export const deleteUser = functions
+  .region('asia-northeast1')
+  .auth.user()
+  .onDelete((user) => {
+    return db.doc(`users/${user.uid}`).delete();
+  });
